+++ src/services/events.ts
@@ -1,0 +1,27 @@
+//	Imports ____________________________________________________________________
+
+
+
+//	Variables __________________________________________________________________
+
+
+
+//	Initialize _________________________________________________________________
+
+
+
+//	Exports ____________________________________________________________________
+
+export * as compare from './events/compare';
+export * as context from './events/context';
+export * as copy from './events/copy';
+export * as deletes from './events/deletes';
+export * as dialogs from './events/dialogs';
+export * as favorites from './events/favorites';
+export * as menu from './events/menu';
+export * as open from './events/open';
+export * as panelstates from './events/panelstates';
+export * as updates from './events/updates';
+
+//	Functions __________________________________________________________________
+