+++ src/services/events/favorites.ts
@@ -1,6 +1,9 @@
 //	Imports ____________________________________________________________________
 
+import { DiffFavoriteMessage } from '../../types';
 
+import { DiffPanel } from '../panel/DiffPanel';
+import { DiffFavorites } from '../sidebar/DiffFavorites';
 
 //	Variables __________________________________________________________________
 
@@ -12,13 +15,11 @@
 
 //	Exports ____________________________________________________________________
 
-export type ContextStates = {
-	[name:string]: boolean,
-};
-
-export type Uri = {
-	fsPath:string,
-};
+export function init (currentDiffPanel:DiffPanel) {
+	
+	currentDiffPanel.msg.on('save:favorite', (data:DiffFavoriteMessage) => DiffFavorites.addFavorite(currentDiffPanel.context, data.pathA, data.pathB));
+	
+}
 
 //	Functions __________________________________________________________________