+++ src/types.ts
@@ -14,6 +14,7 @@
 
 export * from './@types/diffs';
 export * from './@types/extensions';
+export * from './@types/formats';
 export * from './@types/messages';
 export * from './@types/tests';
 
@@ -22,7 +23,6 @@ export * from './services/@types/copy';
 export * from './services/@types/delete';
 export * from './services/@types/events';
 export * from './services/@types/favorites';
-export * from './@types/formats';
 export * from './services/@types/fse';
 export * from './services/@types/history';
 export * from './services/@types/panel';