+++ src/views/components/l13-diff/events/diff.ts
@@ -1,5 +1,7 @@
 //	Imports ____________________________________________________________________
 
+import { DiffInitViewMessage, DiffUpdatePathsMessage } from '../../../../types';
+
 import { msg } from '../../common';
 
 import { L13DiffInputViewModel } from '../../l13-diff-input/l13-diff-input.viewmodel';
@@ -22,25 +24,25 @@ export function init (diff:L13DiffComponent, leftVM:L13DiffInputViewModel, right
 	
 	msg.on('focus', () => setTimeout(() => window.focus(), 0)); // Fixes losing focus if other tab has been closed
 	
-	msg.on('update:paths', (data) => {
+	msg.on('update:paths', (data:DiffUpdatePathsMessage) => {
 		
 		if (data.uris.length) {
-			leftVM.value = (data.uris[0] || 0).fsPath || '';
-			rightVM.value = (data.uris[1] || 0).fsPath || '';
+			leftVM.value = data.uris[0]?.fsPath || '';
+			rightVM.value = data.uris[1]?.fsPath || '';
 			if (data.compare) diff.initCompare();
 		}
 		
 	});
 	
-	msg.on('init:view', (data) => {
+	msg.on('init:view', (data:DiffInitViewMessage) => {
 		
 		msg.removeMessageListener('init:view');
 		
 		diff.initPanelStates(data.panel);
 		
 		if (data.uris.length) {
-			leftVM.value = (data.uris[0] || 0).fsPath || '';
-			rightVM.value = (data.uris[1] || 0).fsPath || '';
+			leftVM.value = data.uris[0]?.fsPath || '';
+			rightVM.value = data.uris[1]?.fsPath || '';
 		} else {
 			leftVM.value = data.workspaces[0] || '';
 			rightVM.value = data.workspaces[1] || '';