+++ src/services/actions/DiffOpen.ts
@@ -1,11 +1,14 @@
 //	Imports ____________________________________________________________________
 
+import { basename } from 'path';
 import * as vscode from 'vscode';
 
 import { Diff, DiffFile } from '../../types';
 
-import { formatNameAndDesc } from '../@l13/formats';
+import { formatName, formatNameAndDesc } from '../@l13/formats';
 import { lstat } from '../@l13/fse';
+
+import * as settings from '../common/settings';
 
 import { SymlinkContentProvider } from './symlinks/SymlinkContentProvider';
 
@@ -89,12 +92,20 @@ export class DiffOpen {
 
 async function openDiff (fileA:DiffFile, fileB:DiffFile, left:vscode.Uri, right:vscode.Uri, openToSide:boolean) {
 	
+	const labelFormat = settings.get('labelFormat');
 	let title = '';
 	
 	if (fileA.name !== fileB.name) {
-		const [label, root] = formatNameAndDesc(fileA.fsPath, fileB.fsPath);
-		title = `${label} (${root})`;
-	} else title = `${fileA.name} (${fileA.root} ↔ ${fileB.root})`;
+		if (labelFormat === 'complete') {
+			const [label, root] = formatNameAndDesc(fileA.fsPath, fileB.fsPath);
+			title = root ? `${label} (${root})` : label;
+		} else if (labelFormat === 'relative') title = `${fileA.name} ↔ ${fileB.name}`;
+		else title = formatName(fileA.fsPath, fileB.fsPath);
+	} else {
+		if (labelFormat === 'complete') title = `${fileA.name} (${fileA.root} ↔ ${fileB.root})`;
+		else if (labelFormat === 'relative') title = fileA.name;
+		else title = basename(fileA.name);
+	}
 	
 	await vscode.commands.executeCommand('vscode.diff', left, right, title, {
 		preview: false,